+++ lib/Doctrine/ODM/PHPCR/UnitOfWork.php
@@ -608,8 +608,8 @@ class UnitOfWork
             $this->documentTranslations[$oid][$locale][$field] = $class->reflFields[$field]->getValue($document);
         }
 
-        if ($this->evm->hasListeners(Event::postLoadTranslation)) {
-            $this->evm->dispatchEvent(Event::postLoadTranslation, new LifecycleEventArgs($document, $this->dm));
+        if ($this->evm->hasListeners(Event::bindTranslation)) {
+            $this->evm->dispatchEvent(Event::bindTranslation, new LifecycleEventArgs($document, $this->dm));
         }
     }